optimized code to dispay outfits using dynamic routing
