--- Client/src/component/outfitcollection/Anarkali.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import './Anarkali.css'
-import Filternavwomen from "../extra component/Filternavwomen";
-
-function Anarkali() {
-    return (
-        <>
-            <main id='outfitmain'>
-                <div>
-                    <Filternavwomen />
-                </div>
-            </main>
-        </>
-    )
-}
-
-export default Anarkali
