--- Client/src/component/outfitcollection/Gown.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import './Gown.css'
-import Filternavwomen from "../extra component/Filternavwomen";
-
-function Gown() {
-    return (
-        <>
-            <main id='outfitmain'>
-                <div>
-                    <Filternavwomen />
-                </div>
-            </main>
-        </>
-    )
-}
-
-export default Gown
