--- Client/src/component/outfitcollection/Lehenga.css
@@ -1,4 +1,0 @@
-#outfitmain {
-    position: relative;
-    top: 16vh;
-}
