--- Client/src/component/outfitcollection/Lehenga.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import './Lehenga.css'
-import Filternavwomen from "../extra component/Filternavwomen";
-
-function Lehenga() {
-    return (
-        <>
-            <main id='outfitmain'>
-                <div>
-                    <Filternavwomen />
-                </div>
-            </main>
-        </>
-    )
-}
-
-export default Lehenga
