--- Client/src/component/outfitcollection/Tuxedo.css
@@ -1,5 +1,0 @@
-#outfitmain {
-    position: relative;
-    top: 16vh;
-}
-
