--- Client/src/component/outfitcollection/Tuxedo.jsx
@@ -1,17 +1,0 @@
-import React from 'react'
-import Filternavmen from "../extra component/Filternavmen";
-import './Tuxedo.css'
-
-function Tuxedo() {
-    return (
-        <>
-            <main id='outfitmain'>
-                <div>
-                    <Filternavmen />
-                </div>
-            </main>
-        </>
-    )
-}
-
-export default Tuxedo
