+++ Client/src/component/outfitcollection/Womenscollection.css
@@ -1,0 +1,11 @@
+#outfitsection2{
+    background: linear-gradient(45deg, rgba(255, 181, 175, 0.5), rgba(173, 46, 36, 0.5));
+    border-radius: 30px;
+    border: 3px solid rgb(232, 171, 108);
+    margin: 3vw;
+    padding: 4vw;
+    display: grid;
+    grid-template-columns: repeat(4, 1fr);
+    gap: 4vw;
+    position: relative;
+}
